+++ src/main.ts
@@ -17,6 +17,15 @@ require('~/main.css');
  */
 
 /**
+ * Font Awesome import
+ */
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { FontAwesomeIcon, FontAwesomeLayers, FontAwesomeLayersText } from '@fortawesome/vue-fontawesome';
+import { faGithub, faTwitter } from '@fortawesome/free-brands-svg-icons';
+
+library.add(faGithub, faTwitter);
+
+/**
  * Layout imports
  */
 import DefaultLayout from '~/layouts/Default.vue';
@@ -26,6 +35,9 @@ import DefaultLayout from '~/layouts/Default.vue';
  */
 const client: ClientApiConstructor = (Vue, { head }) => {
   Vue.component('Layout', DefaultLayout);
+  Vue.component('font-awesome-icon', FontAwesomeIcon);
+  Vue.component('font-awesome-layers', FontAwesomeLayers);
+  Vue.component('font-awesome-layers-text', FontAwesomeLayersText);
   head.title = "My Awesome Gridsome Project";
 };